+++ src/index.ts
@@ -21,7 +21,7 @@ export { default as MainContent } from './components/02-components/main-content/
 export { default as Placeholder } from './components/02-components/placeholder/placeholder';
 export { default as Sidebar } from './components/02-components/sidebar/sidebar';
 export { default as CtaCard } from './components/02-components/cta-card/cta-card';
-export { default as CardExperience } from './components/02-components/card-expericience/card-experience';
+export { default as CardExperience } from './components/02-components/card-experience/card-experience';
 export {
   default as SidebarMenu,
   type SidebarMenuProps,
